+++ common/src/main/kotlin/dev/reqsmith/composer/common/plugin/buildsys/BuildSystem.kt
@@ -24,4 +24,6 @@ interface BuildSystem {
     val resourceFolder: String
 
     fun updateBuildScript(params: MutableMap<String, String>)
+
+    fun build(projectPath: String, buildAndRun: Boolean, infoLogging: Boolean)
 }